+++ graphai/graph.py
@@ -371,15 +371,24 @@ class Graph:
             if current_node.is_end:
                 break
             if current_node.is_router:
-                next_node_name = str(output["choice"])
-                del output["choice"]
-                current_node = self._get_node_by_name(node_name=next_node_name)
-                continue
+                if "choices" in output:
+                    # for parallel execution, we collect next_nodes
+                    choice_names = output["choices"]
+                    del output["choices"]
+                    next_nodes = [self._get_node_by_name(name) for name in choice_names]
+                else:
+                    # otherwise business as usual - get the next node only
+                    next_node_name = str(output["choice"])
+                    del output["choice"]
+                    current_node = self._get_node_by_name(node_name=next_node_name)
+                    continue
+            else:
+                # for sequential execution, we can just get the next node
+                next_nodes = self._get_next_nodes(current_node)
             if stop_at_join and current_node in self.join_nodes:
                 # for parallel branches, wait at JoinEdge until all branches are complete
                 return state
 
-            next_nodes = self._get_next_nodes(current_node)
             if not next_nodes:
                 raise Exception(
                     f"No outgoing edge found for current node '{current_node.name}'."